Got the basis of Ukkonen's algorithm working.
Still missing the unique identifier at the end to transform it to an explicit suffix tree, although this might not be needed for what I am using the suffix tree for?
Need to find out, otherwise the Test enum in suffix_tree/lib.rs will be the basis for this unique separator.