Implemented encoding and decoding for RLZ
Also made a rough quickcheck test to ensure correctness of the encoding and decoding process.
Lastly also made most of the functions accepting &str instead take an AsRef<str> parameter, as that allows for passing either a String or a &str without anything else going on. Only the innermost function had to to call .as_ref() to ensure it is the correct type, and everything else can just pass it along.