Removed the EncodeType to only have to save the substring range.
This would not work if the chosen reference string does not contain a substring, but because this substring can be created from the data, all the characters can be present in the base string by simply appending them.

This removes some overhead in the enum type that rust has, and lead to a good decrease in compressed data for a 50MB file.